[refactor] gemini 리뷰 반영
